pyproject.toml
[tool.poetry]
name = "emfestabloggenerator"
version = "0.1.0"
description = "crew for emfesta blog content generator"
authors = ["Vitor Marques"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "0.28.8"
unstructured = "==0.13.3"
pyowm = "3.3.0"
python-dotenv = "1.0.0"

[tool.poetry.dev-dependencies]

[tool.pyright]
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
